pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teamverse"
version = "0.1.0"
description = "Content generation for PolicyEngine - social images, newsletters, blog posts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "PolicyEngine", email = "hello@policyengine.org"}
]
keywords = ["policyengine", "content", "newsletter", "social-media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "jinja2>=3.0",
    "pillow>=10.0",
    "click>=8.0",
    "httpx>=0.25",
    "fastapi>=0.109",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "black>=23.0",
    "ruff>=0.1",
    "pixelmatch>=0.3",
    "beautifulsoup4>=4.12",
    "html2text>=2020.1.16",
]

[project.scripts]
teamverse = "teamverse.cli:main"

[project.urls]
Homepage = "https://github.com/PolicyEngine/teamverse-py"
Repository = "https://github.com/PolicyEngine/teamverse-py"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=teamverse --cov-report=term-missing"
